tracker: address review — status filter, agent listing, bundler split, secrets test
- single_benchmark: collapse the task-status filter to a single .in_(statuses).
- list_agents: drop the unused subfolder layout; agents are stored as .zip only.
- Split tracker.agent.bundler into bundler (client-side zip packaging) and
  contract (shared yaml/zip contract resolution); update importers.
- Restore the dropped retry/resume secret-override assertion in the CLI test.
