+++ src/valkyrie/cli/bundler.py
@@ -1,10 +1,7 @@
-"""Shim — bundler now lives in tracker.agent_bundler."""
+"""Shim — agent bundling/contract helpers now live in tracker.agent."""
 
-from tracker.agent_bundler import (
-    get_agent_zip_stream,
-    get_contract,
-    get_contract_from_zip_bytes,
-)
+from tracker.agent.bundler import get_agent_zip_stream
+from tracker.agent.contract import get_contract, get_contract_from_zip_bytes
 
 
 __all__ = ["get_agent_zip_stream", "get_contract", "get_contract_from_zip_bytes"]
